apps/ems-api/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
PyJWT==2.8.0
SQLAlchemy==2.0.30
alembic==1.13.1
psycopg[binary]==3.1.19
